+++ corehq/apps/export/views/__init__.py
@@ -27,6 +27,7 @@ from corehq.apps.export.views.list import (
     DeIdDailySavedExportListView,
     DeIdDashboardFeedListView,
     _DeidMixin,
+    download_daily_saved_export,
 )
 from corehq.apps.export.views.new import (
     BaseNewExportView,
